refactor(test): 미션 요구사항 테스트 코드 패키지 분리
