+++ src/test/java/roomescape/mission/MissionStep2Test.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.mission;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
